+++ babel.config.js
@@ -28,6 +28,8 @@ module.exports = function(api) {
                   '@edtr-io/plugin-highlight': '@edtr-io/plugin-highlight/src',
                   '@edtr-io/plugin-hint': '@edtr-io/plugin-hint/src',
                   '@edtr-io/plugin-image': '@edtr-io/plugin-image/src',
+                  '@edtr-io/plugin-important-statement':
+                    '@edtr-io/plugin-important-statement/src',
                   '@edtr-io/plugin-input-exercise':
                     '@edtr-io/plugin-input-exercise/src',
                   '@edtr-io/plugin-rows': '@edtr-io/plugin-rows/src',